mechgen_web/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0,<1
pydantic>=2.7.0,<3
python-dotenv>=1.0.0,<2
google-genai>=0.6.0
Pillow>=10.0.0,<12

-e ../mechgen
